+++ src/main/java/com/example/ms_pedidos/repository/DetallePedidoRepository.java
@@ -3,5 +3,5 @@ package com.example.ms_pedidos.repository;
 import com.example.ms_pedidos.model.DetallePedido;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DetallePedidoRepository extends JpaRepository<DetallePedido, Long> {
+public interface DetallePedidoRepository extends JpaRepository<DetallePedido, Integer> {
 }
